Set Curator Connection Timeout
Curator 5.2.0 has a default "connection timeout" of 15s:

https://github.com/apache/curator/blob/apache-curator-5.2.0/curator-framework/src/main/java/org/apache/curator/framework/CuratorFrameworkFactory.java#L63

And it will throw a warning if the Zookeeper session timeout is less than the Curator connection timeout:

https://github.com/apache/curator/blob/apache-curator-5.2.0/curator-client/src/main/java/org/apache/curator/CuratorZookeeperClient.java#L117-L120

The Hadoop default for ZK timeout is set to 10s:

https://github.com/apache/hadoop/blob/0dd9bf8/hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/CommonConfigurationKeys.java#L414-L416

Which means without setting the "connection timeout" a default installation will keep warning about the connection timeout for Curator is lower than the requested session timeout. This sets both timeout values to override the Curator default.

Another option is to change the default Hadoop ZK timeout from 10s to 15s.